f2fs: remove the unnecessary cast for PTR_ERR
It's not necessary to specify 'int' casting for PTR_ERR.

+++ fs/f2fs/acl.c
@@ -234,7 +234,7 @@ static int f2fs_set_acl(struct inode *inode, int type,
 		value = f2fs_acl_to_disk(acl, &size);
 		if (IS_ERR(value)) {
 			clear_inode_flag(inode, FI_ACL_MODE);
-			return (int)PTR_ERR(value);
+			return PTR_ERR(value);
 		}
 	}
